Refactor motors loader
